requirements.txt
alembic==1.8.1
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
dropbox==11.36.0
filelock==3.8.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-AdminLTE3==1.0.8
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
Flask-Session==0.4.0
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
pdfkit==1.0.0
platformdirs==2.5.2
ply==3.11
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
stone==3.3.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
WTForms==3.0.1
num2words==0.5.12
pandas==1.5.2
Flask-Mail==0.9.1
pillow==9.4.0
babel==7.10.0